Decompiled sub_GAME_7F05E83C

+++ src/game/bondview.h
@@ -216,7 +216,7 @@ struct hand
   s32 field_A7C;
   f32 noise;
   s32 field_A84;
-  s32 field_A88;
+  f32 field_A88;
   s32 field_A8C;
   ObjectRecord* field_A90;
   s32 field_A94;
